Move UIWindow subclass to window.rs
